backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
av==13.1.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloup==3.0.7
Cython==3.1.2
decorator==5.2.1
fastapi==0.115.13
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
isosurfaces==0.1.2
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
networkx==3.5
numpy==2.3.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.4
rich==14.0.0
rsa==4.9.1
scipy==1.15.3
screeninfo==0.8.1
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
srt==3.5.3
starlette==0.46.2
svgelements==1.9.6
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
uvicorn==0.34.3
watchdog==6.0.0
